requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
torch>=1.9.0
torchvision>=0.10.0
pyyaml==6.0.1



